\name{Lab4Assignment-package}
\alias{Lab4Assignment-package}
\alias{Lab4Assignment}
\docType{package}
\title{
\packageTitle{Lab4Assignment}
}
\description{
\packageDescription{Lab4Assignment}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Lab4Assignment}
\packageIndices{Lab4Assignment}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Lab4Assignment}

Maintainer: \packageMaintainer{Lab4Assignment}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
