% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{A linear regression package in R.}
\arguments{
\item{formula, data}{Takes in two parameters, formula takes an expression of formula data type and data takes an dataframe data type.}
}
\value{
Returns the object of linreg class.
}
\description{
This package contains the code for Multiple Regression Model. A linreg function is created which takes two arguments and returns the object with of  RC class linreg.
}

